+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 -   Added new template tag, `responsive_the_title()`, intended to output the page title
     for each single post/page template.
--   Added new filter to the new `responsive_the_title()` template_tag, named `responsive_filter_page_title`
+-   Added new filter to the new `responsive_the_title()` template_tag, named `responsive_filter_the_title`
     which can be used to modify/prevent the output of page titles in responsive-framework.
--   Added additional support feature for bu-banners plugin by changing its banner title markup
+-   Added new functions partial `/inc/bu-banners.php` to provide additional support features for bu-banners
+    plugin. Currently adds the h1 page title to banner if exists and title field is empty, and suppresses
+    the responsive-framework page title.
     to use an h1 element with the current page title if the banner title is left empty.
 -   Refactored `page-templates/calendar.php` for easier child-theming by separating logic
     into functions that now live in the existing `/inc/calendar.php` functions file, and